+++ stac_fastapi/api/stac_fastapi/api/app.py
@@ -12,7 +12,7 @@ from stac_pydantic.api.collections import Collections
 from stac_pydantic.shared import MimeTypes
 from stac_pydantic.version import STAC_VERSION
 from starlette.middleware import Middleware
-from starlette.responses import JSONResponse, Response
+from starlette.responses import Response
 
 from stac_fastapi.api.errors import DEFAULT_STATUS_CODES, add_exception_handlers
 from stac_fastapi.api.middleware import CORSMiddleware, ProxyHeaderMiddleware
@@ -21,8 +21,10 @@ from stac_fastapi.api.models import (
     CollectionUri,
     EmptyRequest,
     GeoJSONResponse,
+    HealthCheck,
     ItemCollectionUri,
     ItemUri,
+    JSONResponse,
 )
 from stac_fastapi.api.openapi import update_openapi
 from stac_fastapi.api.routes import (
@@ -108,7 +110,7 @@ class StacApi:
             ),
             takes_self=True,
         ),
-        converter=update_openapi,
+        converter=update_openapi,  # type: ignore
     )
     router: APIRouter = attr.ib(default=attr.Factory(APIRouter))
     search_get_request_model: Type[BaseSearchGetRequest] = attr.ib(
@@ -396,12 +398,15 @@ class StacApi:
         mgmt_router.add_api_route(
             name="Health",
             path="/_mgmt/health",
-            response_model=Dict,
+            response_model=(
+                HealthCheck if self.settings.enable_response_models else None
+            ),
             responses={
                 200: {
                     "content": {
                         MimeTypes.json.value: {},
                     },
+                    "model": HealthCheck,
                 },
             },
             response_class=self.response_class,