+++ stac_fastapi/extensions/stac_fastapi/extensions/core/collection_search/request.py
@@ -6,8 +6,7 @@ from typing import List, Optional, Tuple, cast
 import attr
 from fastapi import Query
 from pydantic import BaseModel, Field, PrivateAttr, ValidationInfo, field_validator
-from stac_pydantic.api.search import SearchDatetime
-from stac_pydantic.shared import BBox
+from stac_pydantic.shared import BBox, SearchDatetime
 from typing_extensions import Annotated
 
 from stac_fastapi.types.search import (
@@ -24,7 +23,7 @@ from stac_fastapi.types.search import (
 class BaseCollectionSearchGetRequest(APIRequest, DatetimeMixin):
     """Basics additional Collection-Search parameters for the GET request."""
 
-    bbox: Optional[BBox] = attr.ib(default=None, converter=_bbox_converter)
+    bbox: Optional[BBox] = attr.ib(default=None, converter=_bbox_converter)  # type: ignore
     datetime: DateTimeQueryType = attr.ib(default=None, validator=_validate_datetime)
     limit: Annotated[
         Optional[Limit],
@@ -40,21 +39,32 @@ class BaseCollectionSearchPostRequest(BaseModel):
     bbox: Optional[BBox] = Field(
         default=None,
         description="Only return items intersecting this bounding box. Mutually exclusive with **intersects**.",  # noqa: E501
-        openapi_examples={
-            "user-provided": {"value": None},
-            "Montreal": {"value": "-73.896103,45.364690,-73.413734,45.674283"},
+        json_schema_extra={
+            "examples": [
+                # user-provided
+                None,
+                # Montreal
+                "-73.896103,45.364690,-73.413734,45.674283",
+            ],
         },
     )
     datetime: Optional[str] = Field(
         default=None,
         description="""Only return items that have a temporal property that intersects this value.\n
 Either a date-time or an interval, open or closed. Date and time expressions adhere to RFC 3339. Open intervals are expressed using double-dots.""",  # noqa: E501
-        openapi_examples={
-            "user-provided": {"value": None},
-            "datetime": {"value": "2018-02-12T23:20:50Z"},
-            "closed-interval": {"value": "2018-02-12T00:00:00Z/2018-03-18T12:31:12Z"},
-            "open-interval-from": {"value": "2018-02-12T00:00:00Z/.."},
-            "open-interval-to": {"value": "../2018-03-18T12:31:12Z"},
+        json_schema_extra={
+            "examples": [
+                # user-provided
+                None,
+                # single datetime
+                "2018-02-12T23:20:50Z",
+                # closed inverval
+                "2018-02-12T00:00:00Z/2018-03-18T12:31:12Z",
+                # open interval FROM
+                "2018-02-12T00:00:00Z/..",
+                # open interval TO
+                "../2018-03-18T12:31:12Z",
+            ],
         },
     )
     limit: Optional[Limit] = Field(